Revisited all WinMM interop signatures in an attempt to make them safe with x64. (NAudio still builds as x86 only until testing is possible). Nothing seems to have broken in x86 but proceed with caution.
Also, have fixed a longstanding issue with hanging in WaveOutReset due to function callbacks (thanks to Matt69 on the forums)

+++ NAudio/Wave/Compression/AcmInterop.cs
@@ -9,20 +9,22 @@ namespace NAudio.Wave.Compression
     /// </summary>
     class AcmInterop
     {
-        public delegate bool AcmDriverEnumCallback(int hAcmDriverId, int instance, AcmDriverDetailsSupportFlags flags);
+        // http://msdn.microsoft.com/en-us/library/dd742891%28VS.85%29.aspx
+        public delegate bool AcmDriverEnumCallback(IntPtr hAcmDriverId, IntPtr instance, AcmDriverDetailsSupportFlags flags);
 
-        public delegate bool AcmFormatEnumCallback(int hAcmDriverId, ref AcmFormatDetails formatDetails, IntPtr dwInstance, AcmDriverDetailsSupportFlags flags);
+        public delegate bool AcmFormatEnumCallback(IntPtr hAcmDriverId, ref AcmFormatDetails formatDetails, IntPtr dwInstance, AcmDriverDetailsSupportFlags flags);
 
-        public delegate bool AcmFormatTagEnumCallback(int hAcmDriverId, ref AcmFormatTagDetails formatTagDetails, IntPtr dwInstance, AcmDriverDetailsSupportFlags flags);
+        public delegate bool AcmFormatTagEnumCallback(IntPtr hAcmDriverId, ref AcmFormatTagDetails formatTagDetails, IntPtr dwInstance, AcmDriverDetailsSupportFlags flags);
 
         /// <summary>
+        /// http://msdn.microsoft.com/en-us/library/dd742910%28VS.85%29.aspx
         /// UINT ACMFORMATCHOOSEHOOKPROC acmFormatChooseHookProc(
         ///   HWND hwnd,     
         ///   UINT uMsg,     
         ///   WPARAM wParam, 
         ///   LPARAM lParam  
         /// </summary>        
-        public delegate bool AcmFormatChooseHookProc(IntPtr windowHandle, int message, int wParam, int lParam);
+        public delegate bool AcmFormatChooseHookProc(IntPtr windowHandle, int message, IntPtr wParam, IntPtr lParam);
 
         // not done:
         // acmDriverAdd
@@ -40,19 +42,32 @@ namespace NAudio.Wave.Compression
         // acmGetVersion
         // acmStreamMessage
 
+        // http://msdn.microsoft.com/en-us/library/dd742886%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmDriverClose(IntPtr hAcmDriver, int closeFlags);
+        
+        // http://msdn.microsoft.com/en-us/library/dd742890%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
-        public static extern MmResult acmDriverEnum(AcmDriverEnumCallback fnCallback, int dwInstance, AcmDriverEnumFlags flags);
+        public static extern MmResult acmDriverEnum(AcmDriverEnumCallback fnCallback, IntPtr dwInstance, AcmDriverEnumFlags flags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742887%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
-        public static extern MmResult acmDriverDetails(int hAcmDriver, ref AcmDriverDetails driverDetails, int reserved);
+        public static extern MmResult acmDriverDetails(IntPtr hAcmDriver, ref AcmDriverDetails driverDetails, int reserved);
+
+        // http://msdn.microsoft.com/en-us/library/dd742894%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
-        public static extern MmResult acmDriverOpen(out IntPtr pAcmDriver, int hAcmDriverId, int openFlags);
+        public static extern MmResult acmDriverOpen(out IntPtr pAcmDriver, IntPtr hAcmDriverId, int openFlags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742909%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmFormatChoose(ref AcmFormatChoose formatChoose);
+
+        // http://msdn.microsoft.com/en-us/library/dd742914%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmFormatEnum(IntPtr hAcmDriver, ref AcmFormatDetails formatDetails, AcmFormatEnumCallback callback, IntPtr instance, AcmFormatEnumFlags flags);
+        
         /// <summary>
+        /// http://msdn.microsoft.com/en-us/library/dd742916%28VS.85%29.aspx
         /// MMRESULT acmFormatSuggest(
         /// HACMDRIVER had,          
         /// LPWAVEFORMATEX pwfxSrc,  
@@ -77,13 +92,18 @@ namespace NAudio.Wave.Compression
             IntPtr destFormatPointer,
             int sizeDestFormat,
             AcmFormatSuggestFlags suggestFlags);
-        
+
+        // http://msdn.microsoft.com/en-us/library/dd742919%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmFormatTagEnum(IntPtr hAcmDriver, ref AcmFormatTagDetails formatTagDetails, AcmFormatTagEnumCallback callback, IntPtr instance, int reserved);
+
+        // http://msdn.microsoft.com/en-us/library/dd742922%28VS.85%29.aspx
+        // this version of the prototype is for metrics that output a single integer
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmMetrics(IntPtr hAcmObject, AcmMetrics metric, out int output);
         
         /// <summary>
+        /// http://msdn.microsoft.com/en-us/library/dd742928%28VS.85%29.aspx
         /// MMRESULT acmStreamOpen(
         ///   LPHACMSTREAM    phas,       
         ///   HACMDRIVER      had,        
@@ -103,8 +123,8 @@ namespace NAudio.Wave.Compression
             [In, MarshalAs(UnmanagedType.CustomMarshaler, MarshalType = "NAudio.Wave.WaveFormatCustomMarshaler")] 
             WaveFormat destFormat, 
             [In] WaveFilter waveFilter, 
-            int callback, 
-            int instance, 
+            IntPtr callback, 
+            IntPtr instance, 
             AcmStreamOpenFlags openFlags);
 
         /// <summary>
@@ -117,23 +137,32 @@ namespace NAudio.Wave.Compression
             IntPtr sourceFormatPointer,
             IntPtr destFormatPointer,
             [In] WaveFilter waveFilter,
-            int callback,
-            int instance,
+            IntPtr callback,
+            IntPtr instance,
             AcmStreamOpenFlags openFlags);
 
-        
+        // http://msdn.microsoft.com/en-us/library/dd742923%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamClose(IntPtr hAcmStream, int closeFlags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742924%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamConvert(IntPtr hAcmStream, [In, Out] AcmStreamHeaderStruct streamHeader, AcmStreamConvertFlags streamConvertFlags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742929%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamPrepareHeader(IntPtr hAcmStream, [In, Out] AcmStreamHeaderStruct streamHeader, int prepareFlags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742929%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamReset(IntPtr hAcmStream, int resetFlags);
+        
+        // http://msdn.microsoft.com/en-us/library/dd742931%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamSize(IntPtr hAcmStream, int inputBufferSize, out int outputBufferSize, AcmStreamSizeFlags flags);
+
+        // http://msdn.microsoft.com/en-us/library/dd742932%28VS.85%29.aspx
         [DllImport("Msacm32.dll")]
         public static extern MmResult acmStreamUnprepareHeader(IntPtr hAcmStream, [In, Out] AcmStreamHeaderStruct streamHeader, int flags);
-
     }
 }
